--- app/src/pages/playground/PlaygroundApiKeysDropdown.tsx
@@ -1,3 +1,0 @@
-function PlaygroundApiKeysDropdown() {
-  return;
-}
