+++ app/src/store/playground_store/constants.ts
@@ -1,6 +1,7 @@
-enum ChatMessageRole {
-  user = "user",
-  system = "system",
-  ai = "ai",
-  tool = "tool",
-}
+import { LlmProvider } from "./types";
+
+export const ProviderToCredentialDisplayMap: Record<LlmProvider, string> = {
+  openai: "OPENAI_API_KEY",
+  anthropic: "ANTHROPIC_API_KEY",
+  azureOpenAi: "AZURE_OPENAI_API_KEY",
+};
